moving api status to its own file
