+++ src/reactviews/pages/ExecutionPlan/executionPlanInterfaces.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ApiStatus } from "../../../sharedInterfaces/connectionDialog";
+import { ApiStatus } from "../../../sharedInterfaces/webview";
 
 export interface ExecutionPlanWebViewState {
 	sqlPlanContent?: string;
