+++ src/sharedInterfaces/webview.ts
@@ -1,0 +1,11 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export enum ApiStatus {
+	NotStarted = 'notStarted',
+	Loading = 'loading',
+	Loaded = 'loaded',
+	Error = 'error',
+}
